Expanded the example
Integrated DBSingleton
